Got textbox working fine but it is mediocre as of now.
A couple of remarks:
(1) need to generalize it to work at any y-level. (world and battle ones are different)
(2) need to generalize it to work with portraits. (we only have Toriel, could just include her in the spritesheet since she's the only character with one in the ruins).
(3) need to cleanup the source and make it more modular + less messy (change set tilemap calls to a utility function that takes x,y coordinates)
(4) need to animate it for battle transistion where it becomes the place the player's heart goes into

After that it should be good.