Commented out the usage of textbox because I want to PR to master.
Why? I did some calculations on a napkin, there's at max ~1,100 tiles and it takes ~900. I will have to massively optimize it, probably extend it to be like a look-up table. This will require more space, but I think it can be kept <1kb

For now, the changes into state manager and additional states are what are important